simplify some initialization
